Extract light/dark mode toggle, add system preference
